Avoiding warning on PHP log file
From the docs `If set to False and the file already exists, the file
will not be modified even if changes would otherwise be made.
Permissions and ownership will still be enforced, however.`

```
[WARNING ] State for file: /var/log/php_errors.log - Neither 'source'
nor 'contents' nor 'contents_pillar' nor 'contents_grains' was defined,
yet 'replace' was set to 'True'. As there is no source to replace the
file with, 'replace' has been set to 'False' to avoid reading the file
unnecessarily.
```