States for installing proofreader as a Composer global dependency
The requirements for tools (which is what we install as Composer global rather than as a project dependency) become relaxed to allow installation of packages without releases (from dev-master).
There is a new `composer` state that collects all the Composer setup operations to make it easy to require it from other states.