Proposing to access buildvars with `cfg('build_vars.something')`
Build vars are structured like this:
```
{
    "branch": "...",
    "revision": "...",
    "project": {
        "aws": {
        },
        "vagrant": {
        },
    }
    ...
}
```
We currently access them with `cfg('project.revision')` in the Salt
templates or `elife.py` code.

However, this does not give access to the `project` key, but to the root
fo the build vars. I find this confusing as we may have to write `cfg('project.project.aws')` and `project` is used in two places.

So I propose to access them with `cfg('buildvars.revision')` with the
same pattern as `cfg('cfn.DomainName')` used for CloudFormation outputs.
A long backward compatibility window will be in place since there may be
many formulas relying on `project.`