Script for manually setting commit statuses on Github
We shouldn't need to do this in general as Jenkins automates this. However sometimes we have required check on PRs and we want Jenkins to make some automated commits directly on master or other protected branches. Therefore he needs to be able to commit, push and signal that commit as good.