Correct parsing of PHP logs in syslog-ng
We match lines starting with `[` and send as an entry everything until the next line starting with .

Side-effect: it may take a while for a log message to be forwarded if it is the last in the log, as syslog-ng cannot determine if it's finished. In theory it should forward it in 10 seconds, but this doesn't happen.