Extracting backup cron
Should only be run on the first node of a cluster, as we don't need a database to be backed up multiple times in parallel. I assume that if a stack uses multiple servers, these servers don't have local state to backup as they have moved their state in RDS, S3 or a similar external centralized service.

Other crons, like daily-system-updates or time-correction, can (and should) run safely everywhere.