Move custom module call in template file
`elife.image_label` requires docker and some of our scripts to be installed to work. Therefore, it cannot be called inside a  file as that is rendered at the start of the highstate.

Instead, calling the module inside a template for `file.managed` allows all the previous states to run before Docker is used.