Extract uwsgi params
This file can be used by an nginx on the host pointing to a container, so it doesn't need to be tied to installing uwsgi on the host.

It's probably generic enough to be called `wsgi-params` referring to the protocol rather than the daemon; however renaming requires other operations involving nginx so it can't be done at this time.